pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-inbox-agent"
version = "0.1.0"
description = "Daily research paper triage and fast-read agent."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Researcher" }]
dependencies = [
    "pydantic>=2.6",
    "typer>=0.12",
    "httpx>=0.27",
    "feedparser>=6.0",
    "pypdf>=4.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.40"]
pymupdf = ["pymupdf>=1.24"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
paper-inbox = "paper_inbox.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"paper_inbox" = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]
"src/paper_inbox/cli.py" = ["B008"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"
